Create README.md
This is a README document of Stored Procedure I wrote as part of a Snwoflake / Azure Data Factory data pipeline to extract unused customer account records. These accounts are then put into Azure Cloud for the Azure team to remove the accounts from the Dominos system. This is not the final copy of the SP I wrote. But, it is the only version I could find locally because I rarely save client code on my Pc.